Fixed todo's of runtest, changed names of diameter/newdiameter
